requirements.txt
streamlit
together
pandas
matplotlib
seaborn
pdfplumber
python-docx
pytesseract
Pillow
openpyxl